[Refactoring] int type to NSInteger

+++ LYCategory/Classes/_Foundation/_Date/NSDate+Formatter.h
@@ -55,6 +55,6 @@
 
 #pragma mark HOUR MINUTE SECOND
 
-- (int)hourNumberPRC;
+- (NSInteger)hourNumberPRC;
 
 @end
